requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
PySocks==1.7.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==1.26.20
yandex-music==2.2.0
yarl==1.22.0
python-dotenv==1.0.0
